refactor: rename and update use cases and mappers for task state management
