+++ src/main/kotlin/data/datasource/mapper/taskState/TaskStateColumns.kt
@@ -1,6 +1,6 @@
-package org.baghdad.data.datasource.mapper.state
+package data.datasource.mapper.taskState
 
-object StateColumns {
+object TaskStateColumns {
     const val ID = 0
     const val NAME = 1
     const val PROJECT_ID = 2
